# api/docs/openapi_rapidapi_hub.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api-Desk Whales",
    "version": "2.0.0",
    "description": "Multi-chain whale transfer API (threshold ~> USD): live feed, entity labels, risk score and impact estimate. BASIC/PRO: core whale+stats. ULTRA/MEGA: + timeseries and market analysis. Not a generic price/honeypot aggregator."
  },
  "paths": {
...
    "/v1/whales": {
      "get": {
        "tags": [
          "whales"
        ],
        "summary": "List whales",
        "operationId": "listWhales",
        "parameters": [
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset"
            }
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chain"
            }
          },
          {
            "name": "flow_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flow Type"
            }
          },
          {
            "name": "min_usd",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Usd"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhaleListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/whales/tx/{tx_hash}": {
      "get": {
        "tags": [
          "whales"
        ],
        "summary": "Get whale by tx",
        "operationId": "getWhaleByTx",
        "parameters": [
          {
            "name": "tx_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tx Hash"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WhaleEventOut"
                  },
                  "title": "Response Get By Tx V1 Whales Tx  Tx Hash  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/whales/{whale_id}": {
      "get": {
        "tags": [
          "whales"
        ],
        "summary": "Get whale by id",
        "operationId": "getWhale",
        "parameters": [
          {
            "name": "whale_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Whale Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhaleEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/overview": {
      "get": {
        "tags": [
          "whales"
        ],
        "summary": "Stats overview",
        "operationId": "statsOverview",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h",
              "title": "Window"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOverviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/timeseries": {
      "get": {
        "tags": [
          "whales"
        ],
        "summary": "Stats timeseries",
        "operationId": "statsTimeseries",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h",
              "title": "Window"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1h",
              "title": "Bucket"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeseriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entities/{address}": {
      "get": {
        "tags": [
          "whales"
        ],
        "summary": "Get entity",
        "operationId": "getEntity",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ethereum",
              "title": "Chain"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/assets": {
      "get": {
        "tags": [
          "market-signals"
        ],
        "summary": "Market assets",
        "operationId": "marketAssets",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressParty": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": "unknown"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "AddressParty"
      },
...
      "EntityOut": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "chain": {
            "type": "string",
            "title": "Chain"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "meta": {
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "address",
          "chain",
          "label",
          "entity_type",
          "confidence"
        ],
        "title": "EntityOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImpactInfo": {
        "properties": {
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "title": "ImpactInfo"
      },
...
      "RiskInfo": {
        "properties": {
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Factors"
          }
        },
        "type": "object",
        "required": [
          "score",
          "level"
        ],
        "title": "RiskInfo"
      },
...
      "StatsOverviewOut": {
        "properties": {
          "window": {
            "type": "string",
            "title": "Window"
          },
          "total_events": {
            "type": "integer",
            "title": "Total Events"
          },
          "total_volume_usd": {
            "type": "number",
            "title": "Total Volume Usd"
          },
          "by_asset": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Asset"
          },
          "by_flow_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Flow Type"
          },
          "by_chain": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Chain"
          },
          "high_risk_count": {
            "type": "integer",
            "title": "High Risk Count"
          }
        },
        "type": "object",
        "required": [
          "window",
          "total_events",
          "total_volume_usd",
          "by_asset",
          "by_flow_type",
          "by_chain",
          "high_risk_count"
        ],
        "title": "StatsOverviewOut"
      },
      "TimeseriesBucket": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          },
          "volume_usd": {
            "type": "number",
            "title": "Volume Usd"
          },
          "high_risk": {
            "type": "integer",
            "title": "High Risk"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "events",
          "volume_usd",
          "high_risk"
        ],
        "title": "TimeseriesBucket"
      },
      "TimeseriesOut": {
        "properties": {
          "window": {
            "type": "string",
            "title": "Window"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/TimeseriesBucket"
            },
            "type": "array",
            "title": "Buckets"
          }
        },
        "type": "object",
        "required": [
          "window",
          "bucket",
          "buckets"
        ],
        "title": "TimeseriesOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WhaleEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tx_hash": {
            "type": "string",
            "title": "Tx Hash"
          },
          "asset": {
            "type": "string",
            "title": "Asset"
          },
          "chain": {
            "type": "string",
            "title": "Chain"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "amount_usd": {
            "type": "number",
            "title": "Amount Usd"
          },
          "from": {
            "$ref": "#/components/schemas/AddressParty"
          },
          "to": {
            "$ref": "#/components/schemas/AddressParty"
          },
          "flow_type": {
            "type": "string",
            "title": "Flow Type"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskInfo"
          },
          "impact": {
            "$ref": "#/components/schemas/ImpactInfo"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          },
          "block_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Time"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tx_hash",
          "asset",
          "chain",
          "amount",
          "amount_usd",
          "from",
          "to",
          "flow_type",
          "risk",
          "impact",
          "detected_at",
          "provider"
        ],
        "title": "WhaleEventOut"
      },
      "WhaleListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WhaleEventOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count"
        ],
        "title": "WhaleListOut"
      },
...
    }
  },
  "servers": [
    {
      "url": "https://REPLACE_WITH_PUBLIC_API_HOST",
      "description": "Production HTTPS (set after R3 deploy)"
    }
  ]
}